+++ app/test/e2e/queue-status-import-confirmed.spec.js
@@ -496,8 +496,9 @@ describe('STATUS_IMPORT_CONFIRMED handling process', () => {
 
         if (!nock.isDone()) {
             const pendingMocks = nock.pendingMocks();
-            nock.cleanAll();
-            throw new Error(`Not all nock interceptors were used: ${pendingMocks}`);
+            if (pendingMocks.length > 1) {
+                throw new Error(`Not all nock interceptors were used: ${pendingMocks}`);
+            }
         }
 
         await rabbitmqConnection.close();