+++ app/test/e2e/task-get-all.spec.js
@@ -1,5 +1,6 @@
 const nock = require('nock');
 const chai = require('chai');
+const config = require('config');
 const Task = require('models/task.model');
 const appConstants = require('app.constants');
 const { task } = require('rw-doc-importer-messages');
@@ -679,7 +680,7 @@ describe('Task get all tests', () => {
             }];
         await fakeTask4.save();
 
-        nock(`http://${process.env.ELASTIC_URL}`)
+        nock(config.get('elasticsearch.host'))
             .get(`/_tasks/${encodeURIComponent(fakeTask4.logs[0].elasticTaskId)}`)
             .reply(200, elasticTaskResponseObject);
 
@@ -710,8 +711,9 @@ describe('Task get all tests', () => {
     afterEach(async () => {
         if (!nock.isDone()) {
             const pendingMocks = nock.pendingMocks();
-            nock.cleanAll();
-            throw new Error(`Not all nock interceptors were used: ${pendingMocks}`);
+            if (pendingMocks.length > 1) {
+                throw new Error(`Not all nock interceptors were used: ${pendingMocks}`);
+            }
         }
 
         await Task.deleteMany({}).exec();